Acomodando

+++ src/main/java/presenter/SeriesPresenter.java
@@ -6,25 +6,19 @@ import view.ScoredView;
 import view.SearcherView;
 import view.StoredView;
 import view.TVSeriesSearcherWindow;
-
-import javax.swing.*;
 import java.util.ArrayList;
+
+import static utils.TextProcessing.textToHtml;
 
 public class SeriesPresenter {
 
     SearchSeriesModel seriesSearchModel;
-
     SearchWikiPageModel wikiPageModel;
-
     DataBaseModel dataBaseModel;
     SearcherView searchView;
-
     StoredView storedView;
-
     ScoredView scoredView;
-
     Thread taskThread;
-
     TVSeriesSearcherWindow mainWindow;
 
     public SeriesPresenter(SearchSeriesModel seriesSearchModel, SearchWikiPageModel wikiPageModel, DataBaseModel dataBaseModel) {
@@ -36,7 +30,6 @@ public class SeriesPresenter {
 
     public void start(){
         mainWindow = new TVSeriesSearcherWindow(this);
-
         searchView = mainWindow.getSearchView();
         searchView.setUpView();
 
@@ -45,112 +38,89 @@ public class SeriesPresenter {
 
         scoredView = mainWindow.getScoredView();
         scoredView.setUpView();
-        mainWindow.showView();
 
+        mainWindow.showView();
 
         initListeners();
     }
 
     private void initListeners(){
-        seriesSearchModel.addListener(new SearchSeriesModelListener() {
-            @Override public void seriesSearchFinished() {
-                showSeriesSearchResults();
-            }
-        });
+        seriesSearchModel.addListener(this::showSeriesSearchResults);
 
-        wikiPageModel.addListener(new SearchWikiPageModelListener() {
-            @Override public void extractSearchFinished() {
-                showExtractSearchResults();
-            }
-        });
+        wikiPageModel.addListener(this::showExtractSearchResults);
     }
 
     private void showSeriesSearchResults(){
         searchView.showResults(seriesSearchModel.getSearchResults());
-        searchView.setWatingStatus();
+        mainWindow.setWatingStatus();
     }
 
     private void showExtractSearchResults(){
-        searchView.setSearchResultTextPane(wikiPageModel.getExtract() + wikiPageModel.getUrl() );
-
+        searchView.setSearchResultTextPane(wikiPageModel.getExtract() +"<a href='"+wikiPageModel.getUrl()+"'>"+wikiPageModel.getUrl()+"</a>"  );
+        mainWindow.setWatingStatus();
     }
-    public void searchSeries() {
 
+    //SEARCH VIEW SEARCH BUTTON
+    public void searchSeries() {
         taskThread = new Thread(() -> {
-            searchView.setWorkingStatus();
+            mainWindow.setWorkingStatus();
             String seriesName = searchView.getSeriesName();
 
-            //TODO: Controlar vacia
-            seriesSearchModel.searchSeries(seriesName);
-
-
-
+            if(seriesName.isEmpty()) {
+                mainWindow.setWatingStatus();
+                mainWindow.showErrorMessage(searchView.getContentPane(), "Search field is empty!");
+            } else{
+                seriesSearchModel.searchSeries(seriesName);
+            }
         });
         //TODO: TERMINAR HILO???
         taskThread.start();
 
     }
 
-    public void getSelectedExtract(WikiPage selectedResult){
+    public void getSelectedExtract(){
+        taskThread = new Thread(() -> {
+            mainWindow.setWatingStatus();
+            WikiPage selectedResult = searchView.getLastSearchedSeries();
+            getExtract(selectedResult);
+        });
+        taskThread.start();
+    }
+
+    private void getExtract(WikiPage selectedResult){
         wikiPageModel.searchPageExtract(selectedResult);
         if(hasScore(selectedResult.getTitle())) searchScore(selectedResult.getTitle());
         else searchView.showNoScore();
-
     }
 
+    //STORED VIEW
     public void initializeSavedPanel(){
-        Object[] seriesTitles = dataBaseModel.getSavedTitles();
-        Object[] seriesTItlesScored = new Object[seriesTitles.length];
-        for(int i = 0; i < seriesTitles.length; i++){
-            if(hasScore((String) seriesTitles[i])) seriesTItlesScored[i] = seriesTitles[i] + " - Scored";
-            else
-            seriesTItlesScored[i] = seriesTitles[i];
-        }
-        storedView.setSelectSavedComboBox(seriesTItlesScored);
+        storedView.setSelectSavedComboBox(dataBaseModel.getSavedTitles());
     }
 
-    public void showSelectedExtract(){
+    public void getSavedExtract(){
         taskThread = new Thread(() -> {
-            searchView.setWorkingStatus();
-            System.out.println("ENTRE");
+            mainWindow.setWorkingStatus();
             String selectedTitle = storedView.getSeletedSavedTitle();
             String selectedExtract = dataBaseModel.getSavedExtract(selectedTitle);
-            System.out.println(selectedTitle+selectedExtract);
             storedView.setSelectedExtract(textToHtml(selectedExtract));
-            searchView.setWatingStatus();
+            mainWindow.setWatingStatus();
         });
         //TODO: TERMINAR HILO???
         taskThread.start();
-
-
-    }
-    public static String textToHtml(String text) {
-
-        StringBuilder builder = new StringBuilder();
-
-        builder.append("<font face=\"arial\">");
-
-        String fixedText = text
-                .replace("'", "`"); //Replace to avoid SQL errors, we will have to find a workaround..
-
-        builder.append(fixedText);
-
-        builder.append("</font>");
-
-        return builder.toString();
     }
 
     public void deleteSelectedExtract() {
         if(storedView.existSelectedEntry()){
             String selectedTitle = storedView.getSeletedSavedTitle();
+            if(mainWindow.askConfirmation(storedView.getContentPane(), "Are you sure you want to delete the selected entry?")){
+            //TODO: Best con listener o cone xception?
             dataBaseModel.deleteSavedPage(selectedTitle);
             storedView.setSelectSavedComboBox(dataBaseModel.getSavedTitles());
             storedView.emptySavedTextPane();
+            }
         }
     }
-
-
-
     public void recordScore() {
         int score = searchView.getScore();
         WikiPage series = searchView.getLastSearchedSeries();
@@ -175,7 +145,7 @@ public class SeriesPresenter {
     public void onRowSelected() {
         WikiPage selectedSeries = scoredView.getSelectedSeries();
         searchView.setLastSearchedSeries(selectedSeries);
-        getSelectedExtract(selectedSeries);
+        getExtract(selectedSeries);
 
         mainWindow.showSearchPanel();
 
@@ -195,10 +165,10 @@ public class SeriesPresenter {
         try{
             WikiPage pageToSave = searchView.getLastSearchedSeries();
             if(pageToSave != null) updateStoredExtract(pageToSave.getTitle(), pageToSave.getExtract());
-            mainWindow.showSuccessMessage(searchView.getContentPane(), "The series was correctly saved!","Success");
+            mainWindow.showSuccessMessage(searchView.getContentPane(), "The series was correctly saved!");
 
         } catch(Exception e){
-            mainWindow.showErrorMessage(searchView.getContentPane(), e.getMessage(),"Error");
+            mainWindow.showErrorMessage(searchView.getContentPane(), e.getMessage());
 
         }
 
@@ -213,7 +183,7 @@ public class SeriesPresenter {
         int selectedTab = mainWindow.getSelectedTab();
         switch (selectedTab){
             case 0:
-                searchView.setWatingStatus();
+                mainWindow.setWatingStatus();
                 break;
             case 1:
                 initializeSavedPanel();
